requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
blinker==1.7.0
buildtools==1.0.6
causal-conv1d @ file:///home/ju/Desktop/NetMamba/causal_conv1d-1.1.1%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=f572458de281ed15fbd6f9262d44d66591416c7a903c4da699ed4d70b3cf5a83
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
community==1.0.0b1
constantly==23.10.4
contourpy==1.2.0
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.14.1
dotenv==0.9.9
dpkt==1.9.8
easydict==1.13
einops==0.7.0
entrypoints==0.4
-e git+https://github.com/alibaba/EssentialMC2.git@3e5de2cf88ab279617a460e934ea343cb7040999#egg=essmc2
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
filelock==3.13.1
Flask==3.0.0
flowcontainer==7.4
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
grpcio==1.62.1
gunicorn==21.2.0
huggingface-hub==0.19.4
hyperlink==21.0.0
hyperopt==0.2.7
idna==3.6
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==7.0.0
incremental==22.10.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.0
mamba-ssm @ file:///home/ju/Desktop/NetMamba/mamba_ssm-1.1.3.post1%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=4470e65acf34eb9f3d350546c18e20d8af3d002290109a95abea049145dd2bb9
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-transforms==0.0.6
orderedmultidict==1.0.1
oss2==2.19.1
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.9.0
progress==1.6
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==24.9.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflowmeter==0.2.4
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-hostlist==1.23.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
randaugment==1.0.2
redo==2.0.4
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scapy==2.5.0
scikit-learn==1.3.2
scikit-optimize==0.10.2
scipy==1.11.3
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
torch @ file:///home/ju/Desktop/NetMamba/torch-2.1.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=ec76d11350c8e887a34d36602cd37f50639bc9cda9faea4d43f2070e9792e4a4
torchaudio==2.1.1
torchprofile==0.0.4
torchstat==0.0.7
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.35.2
triton==2.1.0
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
wandb==0.19.6
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xgboost==2.1.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
